requirements.txt
tensorflow==2.17.1
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.9.0.80
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
werkzeug==2.3.7
